update within major

+++ apps/squzy_incident/expression/snapshot.go
@@ -118,11 +118,11 @@ func (e *expressionStruct) getSnapshotEnv(schedulerId string) map[string]interfa
 		"getValue": func(snapshot *apiPb.SchedulerSnapshot) *structpb.Value {
 			return snapshot.GetMeta().GetValue()
 		},
-		"unixToTime": func(unix int64) time.Time {
-			return time.Unix(unix, 0)
+		"unixToTime": func(unix int) time.Time {
+			return time.Unix(int64(unix), 0)
 		},
-		"unixNanoToTime": func(unixNano int64) time.Time {
-			return time.Unix(0, unixNano)
+		"unixNanoToTime": func(unixNano int) time.Time {
+			return time.Unix(0, int64(unixNano))
 		},
 		"null": nil,
 		"mulDuration": func(f int, t time.Duration) time.Duration {
